+++ jni/tests/faiss_wrapper_test.cpp
@@ -30,17 +30,14 @@ TEST(FaissCreateIndexTest, BasicAssertions) {
     // Define the data
     faiss::idx_t numIds = 200;
     std::vector<faiss::idx_t> ids;
-    std::vector<std::vector<float>> vectors;
+    auto *vectors = new std::vector<float>();
     int dim = 2;
+    vectors->reserve(dim * numIds);
     for (int64_t i = 0; i < numIds; ++i) {
         ids.push_back(i);
-
-        std::vector<float> vect;
-        vect.reserve(dim);
         for (int j = 0; j < dim; ++j) {
-            vect.push_back(test_util::RandomFloat(-500.0, 500.0));
+            vectors->push_back(test_util::RandomFloat(-500.0, 500.0));
         }
-        vectors.push_back(vect);
     }
 
     std::string indexPath = test_util::RandomString(10, "tmp/", ".faiss");
@@ -58,12 +55,12 @@ TEST(FaissCreateIndexTest, BasicAssertions) {
     EXPECT_CALL(mockJNIUtil,
                 GetJavaObjectArrayLength(
                         jniEnv, reinterpret_cast<jobjectArray>(&vectors)))
-            .WillRepeatedly(Return(vectors.size()));
+            .WillRepeatedly(Return(vectors->size()));
 
     // Create the index
     knn_jni::faiss_wrapper::CreateIndex(
             &mockJNIUtil, jniEnv, reinterpret_cast<jintArray>(&ids),
-            reinterpret_cast<jobjectArray>(&vectors), (jstring)&indexPath,
+            (jlong) vectors, dim , (jstring)&indexPath,
             (jobject)&parametersMap);
 
     // Make sure index can be loaded
@@ -77,17 +74,14 @@ TEST(FaissCreateIndexFromTemplateTest, BasicAssertions) {
     // Define the data
     faiss::idx_t numIds = 100;
     std::vector<faiss::idx_t> ids;
-    std::vector<std::vector<float>> vectors;
+    auto *vectors = new std::vector<float>();
     int dim = 2;
+    vectors->reserve(dim * numIds);
     for (int64_t i = 0; i < numIds; ++i) {
         ids.push_back(i);
-
-        std::vector<float> vect;
-        vect.reserve(dim);
         for (int j = 0; j < dim; ++j) {
-            vect.push_back(test_util::RandomFloat(-500.0, 500.0));
+            vectors->push_back(test_util::RandomFloat(-500.0, 500.0));
         }
-        vectors.push_back(vect);
     }
 
     std::string indexPath = test_util::RandomString(10, "tmp/", ".faiss");
@@ -105,7 +99,7 @@ TEST(FaissCreateIndexFromTemplateTest, BasicAssertions) {
     EXPECT_CALL(mockJNIUtil,
                 GetJavaObjectArrayLength(
                         jniEnv, reinterpret_cast<jobjectArray>(&vectors)))
-            .WillRepeatedly(Return(vectors.size()));
+            .WillRepeatedly(Return(vectors->size()));
 
     std::string spaceType = knn_jni::L2;
     std::unordered_map<std::string, jobject> parametersMap;
@@ -113,7 +107,7 @@ TEST(FaissCreateIndexFromTemplateTest, BasicAssertions) {
 
     knn_jni::faiss_wrapper::CreateIndexFromTemplate(
             &mockJNIUtil, jniEnv, reinterpret_cast<jintArray>(&ids),
-            reinterpret_cast<jobjectArray>(&vectors), (jstring)&indexPath,
+            (jlong)vectors, dim, (jstring)&indexPath,
             reinterpret_cast<jbyteArray>(&(vectorIoWriter.data)),
             (jobject) &parametersMap
             );
@@ -486,17 +480,14 @@ TEST(FaissCreateHnswSQfp16IndexTest, BasicAssertions) {
     // Define the data
     faiss::idx_t numIds = 200;
     std::vector<faiss::idx_t> ids;
-    std::vector<std::vector<float>> vectors;
+    auto *vectors = new std::vector<float>();
     int dim = 2;
+    vectors->reserve(dim * numIds);
     for (int64_t i = 0; i < numIds; ++i) {
         ids.push_back(i);
-
-        std::vector<float> vect;
-        vect.reserve(dim);
         for (int j = 0; j < dim; ++j) {
-            vect.push_back(test_util::RandomFloat(-500.0, 500.0));
+            vectors->push_back(test_util::RandomFloat(-500.0, 500.0));
         }
-        vectors.push_back(vect);
     }
 
     std::string indexPath = test_util::RandomString(10, "tmp/", ".faiss");
@@ -514,12 +505,12 @@ TEST(FaissCreateHnswSQfp16IndexTest, BasicAssertions) {
     EXPECT_CALL(mockJNIUtil,
                 GetJavaObjectArrayLength(
                         jniEnv, reinterpret_cast<jobjectArray>(&vectors)))
-            .WillRepeatedly(Return(vectors.size()));
+            .WillRepeatedly(Return(vectors->size()));
 
     // Create the index
     knn_jni::faiss_wrapper::CreateIndex(
             &mockJNIUtil, jniEnv, reinterpret_cast<jintArray>(&ids),
-            reinterpret_cast<jobjectArray>(&vectors), (jstring)&indexPath,
+            (jlong)vectors, dim, (jstring)&indexPath,
             (jobject)&parametersMap);
 
     // Make sure index can be loaded