+++ src/main/java/org/opensearch/knn/jni/FaissService.java
@@ -50,27 +50,33 @@ class FaissService {
     }
 
     /**
-     * Create an index for the native library
+     * Create an index for the native library The memory occupied by the vectorsAddress will be freed up during the
+     * function call. So Java layer doesn't need to free up the memory. This is not an ideal behavior because Java layer
+     * created the memory address and that should only free up the memory. We are tracking the proper fix for this on this
+     * <a href="https://github.com/opensearch-project/k-NN/issues/1600">issue</a>
      *
      * @param ids array of ids mapping to the data passed in
-     * @param data array of float arrays to be indexed
+     * @param vectorsAddress address of native memory where vectors are stored
+     * @param dim dimension of the vector to be indexed
      * @param indexPath path to save index file to
      * @param parameters parameters to build index
      */
-    public static native void createIndex(int[] ids, float[][] data, String indexPath, Map<String, Object> parameters);
+    public static native void createIndex(int[] ids, long vectorsAddress, int dim, String indexPath, Map<String, Object> parameters);
 
     /**
      * Create an index for the native library with a provided template index
      *
      * @param ids array of ids mapping to the data passed in
-     * @param data array of float arrays to be indexed
+     * @param vectorsAddress address of native memory where vectors are stored
+     * @param dim dimension of the vector to be indexed
      * @param indexPath path to save index file to
      * @param templateIndex empty template index
      * @param parameters additional build time parameters
      */
     public static native void createIndexFromTemplate(
         int[] ids,
-        float[][] data,
+        long vectorsAddress,
+        int dim,
         String indexPath,
         byte[] templateIndex,
         Map<String, Object> parameters
@@ -173,33 +179,15 @@ class FaissService {
     public static native byte[] trainIndex(Map<String, Object> indexParameters, int dimension, long trainVectorsPointer);
 
     /**
+     * <p>
+     * The function is deprecated. Use {@link JNICommons#storeVectorData(long, float[][], long)}
+     * </p>
      * Transfer vectors from Java to native
      *
      * @param vectorsPointer pointer to vectors in native memory. Should be 0 to create vector as well
      * @param trainingData data to be transferred
      * @return pointer to native memory location of training data
      */
+    @Deprecated(since = "2.14.0", forRemoval = true)
     public static native long transferVectors(long vectorsPointer, float[][] trainingData);
-
-    /**
-     * Transfer vectors from Java to native layer. This is the version 2 of transfer vector functionality. The
-     * difference between this and the version 1 is, this version puts vectors at the end rather than in front.
-     * Keeping this name as V2 for now, will come up with better name going forward.
-     * <p>
-     * TODO: Rename the function
-     * <br>
-     * TODO: Make this function native function and use a common cpp file to host these functions.
-     * </p>
-     * @param vectorsPointer pointer to vectors in native memory. Should be 0 to create vector as well
-     * @param data data to be transferred
-     * @return pointer to native memory location for data
-     */
-    public static native long transferVectorsV2(long vectorsPointer, float[][] data);
-
-    /**
-     * Free vectors from memory
-     *
-     * @param vectorsPointer to be freed
-     */
-    public static native void freeVectors(long vectorsPointer);
 }