+++ src/testFixtures/java/org/opensearch/knn/TestUtils.java
@@ -6,17 +6,20 @@
 package org.opensearch.knn;
 
 import com.google.common.collect.ImmutableMap;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
 import org.opensearch.common.xcontent.XContentHelper;
 import org.opensearch.core.common.bytes.BytesArray;
 import org.opensearch.core.xcontent.DeprecationHandler;
 import org.opensearch.core.xcontent.NamedXContentRegistry;
 import org.opensearch.core.xcontent.MediaTypeRegistry;
-import org.opensearch.knn.index.codec.util.KNNCodecUtil;
 import java.io.BufferedReader;
 import java.io.FileReader;
 import java.io.IOException;
 import org.opensearch.knn.index.SpaceType;
 import org.opensearch.knn.index.codec.util.SerializationMode;
+import org.opensearch.knn.jni.JNICommons;
 import org.opensearch.knn.plugin.script.KNNScoringUtil;
 import java.util.Comparator;
 import java.util.Random;
@@ -248,7 +251,7 @@ public class TestUtils {
      * Class to read in some test data from text files
      */
     public static class TestData {
-        public KNNCodecUtil.Pair indexData;
+        public Pair indexData;
         public float[][] queries;
 
         public TestData(String testIndexVectorsPath, String testQueriesPath) throws IOException {
@@ -256,7 +259,7 @@ public class TestUtils {
             queries = readQueries(testQueriesPath);
         }
 
-        private KNNCodecUtil.Pair readIndexData(String path) throws IOException {
+        private Pair readIndexData(String path) throws IOException {
             List<Integer> idsList = new ArrayList<>();
             List<Float[]> vectorsList = new ArrayList<>();
 
@@ -292,8 +295,7 @@ public class TestUtils {
                     vectorsArray[i][j] = vectorsList.get(i)[j];
                 }
             }
-
-            return new KNNCodecUtil.Pair(idsArray, vectorsArray, SerializationMode.COLLECTION_OF_FLOATS);
+            return new Pair(idsArray, vectorsArray[0].length, SerializationMode.COLLECTION_OF_FLOATS, vectorsArray);
         }
 
         private float[][] readQueries(String path) throws IOException {
@@ -323,6 +325,20 @@ public class TestUtils {
                 }
             }
             return queryArray;
+        }
+
+        public long loadDataToMemoryAddress() {
+            return JNICommons.storeVectorData(0, indexData.vectors, (long) indexData.vectors.length * indexData.vectors[0].length);
+        }
+
+        @AllArgsConstructor
+        public static class Pair {
+            public int[] docs;
+            @Getter
+            @Setter
+            private int dimension;
+            public SerializationMode serializationMode;
+            public float[][] vectors;
         }
     }
 }